Mapping study table view with backend
